Don't destroy depleted batteries or ammo if multiple exists

+++ src/module/item/ammo/document.ts
@@ -3,6 +3,7 @@ import type { RawItemChatData } from "@item/base/data/index.ts";
 import type { ConsumableTrait } from "@item/consumable/types.ts";
 import { PhysicalItemPF2e } from "@item/physical/index.ts";
 import type { WeaponPF2e } from "@item/weapon/document.ts";
+import { getLoadedAmmo } from "@item/weapon/helpers.ts";
 import { PickAThingPrompt } from "@module/apps/pick-a-thing-prompt/app.ts";
 import type { ValueAndMax } from "@module/data.ts";
 import type { EnrichmentOptionsPF2e } from "@system/text-editor.ts";
@@ -114,26 +115,35 @@ class AmmoPF2e<TParent extends ActorPF2e | null = ActorPF2e | null> extends Phys
             ChatMessage.create({ speaker, content, flags });
         }
 
-        // Optionally destroy the item or deduct charges or quantity
-        // Also keep if it has rule elements
-        if (this.system.uses.autoDestroy && uses.value <= thisMany) {
-            const newQuantity = Math.max(this.quantity - (uses.max > 1 ? 1 : thisMany), 0);
-            const isPreservedAmmo = this.system.rules.length > 0;
-            if (newQuantity <= 0 && !isPreservedAmmo) {
-                await this.delete();
-            } else {
-                await this.update({
-                    "system.quantity": newQuantity,
-                    "system.uses.value": uses.max,
-                });
-            }
-        } else if (!this.system.uses.autoDestroy && this.parentItem && uses.value <= thisMany) {
-            await this.update({ "system.uses.value": this.system.uses.max }, { render: false });
-            await this.detach({ quantity: thisMany, skipConfirm: true });
+        // If we have more uses than we need to spend, update uses and return early
+        if (uses.value > thisMany) {
+            await this.update({ "system.uses.value": Math.max(uses.value - thisMany, 0) });
+            return;
+        }
+
+        const autoDestroy = this.system.uses.autoDestroy;
+        const newQuantity = Math.max(this.quantity - (uses.max > 1 ? 1 : thisMany), 0);
+
+        // Check if we need to preserve 0 quantity ammo for rule element reasons,
+        // either because the ammo has it or the next in line might have some.
+        const loadedAmmo = this.parentItem?.isOfType("weapon")
+            ? getLoadedAmmo(this.parentItem).filter((a) => a.quantity > 0)
+            : null;
+        const isPreservedAmmo = this.system.rules.length > 0 || (loadedAmmo?.length ?? 0) > 1;
+
+        if (autoDestroy && newQuantity <= 0 && !isPreservedAmmo) {
+            // Delete ammo if this is the last one and we don't care about preserving it
+            await this.delete();
+        } else if (this.parentItem && !this.isMagazine && !autoDestroy) {
+            // Detach loaded non-destroyable ammo
+            const numDetach = this.quantity - newQuantity;
+            await this.detach({ quantity: numDetach, keepZero: isPreservedAmmo, skipConfirm: true });
+        } else if (this.parentItem && this.isMagazine) {
+            // Internal non-destroyable magazines (batteries) keep 1 quantity and 0 uses
+            await this.update({ "system.quantity": 1, "system.uses.value": 0 });
         } else {
-            await this.update({
-                "system.uses.value": Math.max(uses.value - thisMany, 0),
-            });
+            // Update quantity to new value, and roll over uses in case this is non-loaded magazine ammo
+            await this.update({ "system.quantity": newQuantity, "system.uses.value": uses.max });
         }
     }
